Better encapsulation / breakdown
